Give up
Just spent waaaay too long trying to get well-behaved output from
this (hmmer) method of partitioning. Basic problem is that it isn't
designed with what we really have in mind, so all these hacks have to
happen on top of what hmmer thinks it's supposed to do. It's too complicated
to go into here, but the final straw was the clusterfuck of trying to
eliminated matches with indels in the interior of the sequence. I don't
want those! There's prolly some way to turn 'em off, but the last straw
means you shoulda stopped a while ago, eh idiot?